refactor: moved classes around for consistency
