+++ kinotic-domain/src/main/java/org/kinotic/domain/internal/api/rest/mcp/model/JsonRpcResponse.java
@@ -1,4 +1,4 @@
-package org.kinotic.gateway.internal.mcp.model;
+package org.kinotic.domain.internal.api.rest.mcp.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Getter;
